지도의 마커 정보 DB 조회, 추가 기능 구현
프론트엔드
- 클릭한 마커, DB에 저장된 마커의 창을 분리하여 html 문자열 생성
- DB에 저장된 마커를 클릭한 마커와 다른 아이콘 사용하도록 변경
- 서버에서 Member의 마커 리스트들을 json으로 받아서 지도요 마크 생성
백엔드
- 프론트에서 get메소드로 보낸 위도, 경도 정보의 마커를 DB에 저장

+++ mapBookmarkSite/mapbookmark/models.py
@@ -8,7 +8,7 @@ class Member(models.Model):
 
 class Marker(models.Model):
     member_id = models.ForeignKey(Member, on_delete=models.CASCADE)
-    latitude = models.IntegerField(default=0)
-    longitude = models.IntegerField(default=0)
+    latitude = models.FloatField(default=0.0)
+    longitude = models.FloatField(default=0.0)
     title = models.CharField(max_length=100)
     content = models.CharField(max_length=300)
